requirements.txt
google-generativeai
python-dotenv==1.0.1
scikit-learn==1.5.1
evaluate==0.4.3
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.6
langchain-experimental==0.3.2
langchain-google-genai==2.0.0
langchain-openai==0.2.1
langchain-text-splitters==0.3.0
llama-index==0.11.14
matplotlib==3.9.2
seaborn==0.13.2
pandas==2.2.2
numpy==1.26.4
nltk==3.9.1
sacrebleu==2.4.3
bert-score==0.3.13
requests==2.32.3
uvicorn==0.30.6
flask
rouge_score
sentence-transformers==3.2.0